refactor(test) : 테스트에 활용되는 유틸 함수 이동
- 유틸 함수로 활용하기 위해 파일을 분리하였으나 분리한 파일도 테스트 파일로 인식하여 테스트에 포함되는 현상 발생
- jest.config.js를 통해 설정을 바꾸려했으나, package.json 파일 내에 존재하는 jest 세팅 값과 충돌하여 함수를 ApplicationTest로 이동하여 export
