--- __tests__/testUtils.js
@@ -1,11 +1,0 @@
-import { MissionUtils } from '@woowacourse/mission-utils';
-
-export const mockQuestions = (inputs) => {
-  MissionUtils.Console.readLineAsync = jest.fn();
-
-  MissionUtils.Console.readLineAsync.mockImplementation(() => {
-    const input = inputs.shift();
-
-    return Promise.resolve(input);
-  });
-};
